+++ src/extension.ts
@@ -4,7 +4,7 @@ import { KeymapParser } from './Parser';
 import { checkForIncorrectFileAssociations } from './settings';
 
 export async function activate(context: vscode.ExtensionContext) {
-    const parser = await KeymapParser.init();
+    const parser = await KeymapParser.init(context);
     context.subscriptions.push(parser, new KeymapAnalyzer(parser));
 
     checkForIncorrectFileAssociations();