+++ webpack.config.js
@@ -1,0 +1,115 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+//@ts-check
+'use strict';
+
+//@ts-check
+/** @typedef {import('webpack').Configuration} WebpackConfig **/
+
+const path = require('path');
+const webpack = require('webpack');
+
+/** @type WebpackConfig */
+const webExtensionConfig = {
+    mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
+    target: 'webworker', // extensions run in a webworker context
+    entry: {
+        extension: './src/extension.ts',
+        'test/suite/index': './src/test/suite/index.ts',
+    },
+    output: {
+        filename: '[name].js',
+        path: path.join(__dirname, './dist/web'),
+        libraryTarget: 'commonjs',
+        devtoolModuleFilenameTemplate: '../../[resource-path]',
+    },
+    resolve: {
+        mainFields: ['browser', 'module', 'main'], // look for `browser` entry point in imported node modules
+        extensions: ['.ts', '.js'],
+        alias: {
+            // provides alternate implementation for node module and source files
+            './wasm': path.join(__dirname, './src/web/wasm.ts'),
+        },
+        fallback: {
+            // Webpack 5 no longer polyfills Node.js core modules automatically.
+            // see https://webpack.js.org/configuration/resolve/#resolvefallback
+            // for the list of Node.js core module polyfills.
+            assert: require.resolve('assert'),
+            // web-tree-sitter tries to import "fs", which can be ignored.
+            // https://github.com/tree-sitter/tree-sitter/issues/466
+            fs: false,
+            path: false,
+        },
+    },
+    module: {
+        rules: [
+            {
+                test: /\.ts$/,
+                exclude: /node_modules/,
+                use: [
+                    {
+                        loader: 'ts-loader',
+                    },
+                ],
+            },
+        ],
+    },
+    plugins: [
+        new webpack.ProvidePlugin({
+            process: 'process/browser', // provide a shim for the global `process` variable
+        }),
+    ],
+    externals: {
+        vscode: 'commonjs vscode', // ignored because it doesn't exist
+    },
+    performance: {
+        hints: false,
+    },
+    devtool: 'nosources-source-map', // create a source map that points to the original source file
+    infrastructureLogging: {
+        level: 'log', // enables logging required for problem matchers
+    },
+};
+
+/** @type WebpackConfig */
+const extensionConfig = {
+    mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
+    target: 'node', // vscode extensions run in a Node.js-context
+
+    entry: './src/extension.ts',
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'extension.js',
+        libraryTarget: 'commonjs2',
+    },
+    externals: {
+        vscode: 'commonjs vscode', // the vscode-module is created on-the-fly and must be excluded.
+        // Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
+        // modules added here also need to be added in the .vscodeignore file
+    },
+    resolve: {
+        extensions: ['.ts', '.js'],
+    },
+    module: {
+        rules: [
+            {
+                test: /\.ts$/,
+                exclude: /node_modules/,
+                use: [
+                    {
+                        loader: 'ts-loader',
+                    },
+                ],
+            },
+        ],
+    },
+    devtool: 'nosources-source-map', // create a source map that points to the original source file
+    infrastructureLogging: {
+        level: 'log', // enables logging required for problem matchers
+    },
+};
+
+module.exports = [extensionConfig, webExtensionConfig];